move timeline to separate area
