+++ src/lib/components/Region.svelte
@@ -1,10 +1,14 @@
 <script lang="ts">
-  let that: HTMLElement | undefined = undefined;
+  export let el: HTMLElement | undefined = undefined;
   let clazz = "";
 
-  export { that as this, clazz as class };
+  export { clazz as class };
 </script>
 
-<article class="border border-gray-300 rounded-lg p-2 {clazz}" bind:this={that}>
+<article
+  class="border border-gray-300 rounded-lg p-2 {clazz}"
+  bind:this={el}
+  on:wheel
+>
   <slot />
 </article>
